Added button hover system & updated README

+++ include/SCGUI/types.h
@@ -40,12 +40,12 @@ struct SCGUI_Color {
 };
 
 struct SCGUI_Vector2 {
-	double x;
-	double y;
+	int x;
+	int y;
 };
 
 struct SCGUI_Color new_color(uint8_t red, uint8_t green, uint8_t blue);
-struct SCGUI_Vector2 new_vec2(double x, double y);
+struct SCGUI_Vector2 new_vec2(int x, int y);
 
 
 #endif
